+++ test/internal/unit/traces/traceHandler.test.ts
@@ -109,6 +109,7 @@ describe("Library/TraceHandler", () => {
       _config.instrumentationOptions = {
         http: { enabled: false },
         azureSdk: { enabled: false },
+        azureFunctions: { enabled: false },
         mongoDb: { enabled: false },
         mySql: { enabled: false },
         postgreSql: { enabled: false },
@@ -258,6 +259,7 @@ describe("Library/TraceHandler", () => {
       _config.instrumentationOptions = {
         http: { enabled: true },
         azureSdk: { enabled: false },
+        azureFunctions: { enabled: false },
         mongoDb: { enabled: false },
         mySql: { enabled: false },
         postgreSql: { enabled: false },